// models/student-data.ts
export interface StudentData {
    scenarioId: string;
    firstName: string;
    lastName: string;
    email: string;
    gender: string;
    mobile: string;
    dateOfBirth: string;
    subjects: string[];
    hobbies: string[];
    pictureFileName: string;
    currentAddress: string;
    state: string;
    city: string;
}